Update CMake configuration and enhance async logging functionality; refactor various components for improved performance and readability

+++ include/Interface/IBaseCombatCharacter.h
@@ -79,7 +79,7 @@ public:
 
 	bool IsMarkSet()
 	{
-		return ( m_flMarkTolerance != NO_MARK );
+		return ( m_flMarkTolerance != static_cast<float>(NO_MARK) );
 	}
 
 	bool TargetMoved( IBaseEntity *pEntity )
